LayoutTests/http/tests/notifications/request-expected.txt

This test checks that a request for permission is made.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "default"
PASS permission is granted.
PASS Notification.permission is "granted"
PASS successfullyParsed is true

TEST COMPLETE
